agent/requirements.txt
openai>=1.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
psycopg2-binary>=2.9.9
redis>=5.0.0
pgvector>=0.3.5
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
sentence-transformers>=3.0.0
minio>=7.2.0
python-multipart>=0.0.9
